# Performance: Avoid calling ParseDuplicatePolicy() on add() with argv=NULL